requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.2
bleach==6.2.0
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.1
django-js-asset==3.1.2
django-resized==1.0.3
django-tinymce==4.1.0
et_xmlfile==2.0.0
idna==3.10
lxml==4.9.3
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.31.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
